Combine use case with service
